Build out CRUD API
Update the README file inherited from TKOut to include additional CRUD
operations for reading all of a restaurants, creating a new popular dish
for a specified restaurant, update the properties of a specific dish for
a specific restaurant, and removing a specific popular dish from a
specific restaurant. Also, update the links to my team member's services
repos they will be working on.