Finish generating data and qork on testing Queries
Generate the final set of data to use for both PostgresQL and Cassandra.
Configure my cassandra schema to be compatible with the queries I need
to make. Run timed CRUD operations on both databases to gain understanding of
where to optimize and gather information to choose which database to use
moving forward.